congestion: multi-hop transactions in all-for-one workload (#10762)
Rather than just sending single-receipt actions, we want transactions
that touch multiple shards. Specifically, we want workloads that are not
very predictable with regards to which shard will be burning most of the
gas.

With this commit, we have an "All To One" and an "Indirect All To One"
strategy. The first is what we had before, depth 1 transactions only.
The second is with transaction of up to depth 3.