rpc: return timeout error from tx_status_fetch (#10789)
All the logic of `tx_status_fetch` function is: we poll tx_status method
and wait for the desired level of tx finality.
`tx_status_fetch` is used in several places including
`broadcast_tx_commit` RPC method.

With the chunk congestions we have right now, we return
`UNKNOWN_TRANSACTION` error to `broadcast_tx_commit` after 20 seconds of
waiting, which is both sad and weird.

The error we store in `tx_status_result` is not good enough to show it
to the user, otherwise we would break from the loop with it immediately.
If we reach the timeout boundary, I suggest always to return timeout
error.