yield/resume: directly process receipts from the timeout queue (#10800)
**What?**

Before: when a yield timed out, we would resolve it by creating a new
outgoing data receipt.
After: we directly process the data receipt, which we know for a fact is
destined for the local shard.

To keep the balance checker happy, these receipts are essentially
interpreted as additional incoming receipts.

**Why?**

This is both an efficiency improvement and a correctness fix. We avoid
the need to create an outgoing receipt. We also eliminate the following
possibility:

1. User submits a transaction which creates a yielded promise.
2. After some time, the timeout triggers and the protocol delivers a new
outgoing receipt. Due to congestion, the timeout receipt enters the
delayed receipts queue.
3. The user submits a transaction which calls yield resume. The yield
callback hasn't executed, so the user gets back a success signal
indicating the yield payload was submitted successfully. When returning
such a signal we guarantee that the callback will be executed with a
user-submitted payload.
4. The timeout receipt is delivered and the yield callback is executed
without a payload. The user is sad and confused.

After this change, the delayed receipts queue must be empty for a
timeout to trigger. When the timeout triggers, the yield callback will
be executed immediately. It eliminates such timing issues.