congestion: Max size, gas and bytes in summary (#10793)
Instead of showing the snapshot of the model in the end, it is better to
display the max values throughout the model execution in the final
summary table.

Also add gas and byte values of queues, rather than just the number of
receipts.

This makes the summary table much more useful to quickly compare the
performance of many different strategies.