yield/resume: gas fee estimations (#10764)
Somewhat surprising to me is that estimations for the resume are much
more high than for create. I suspect that the estimation is covering a
significant part of the transaction runtime that is not really related
to the resumes, but I'm struggling to isolate the interesting parts
well.

So for the time being I'm submitting the baseline safe estimates. In
practice the time will be (likely, much) lower than the fees charged for
the operations, but the fees are still low enough that these host
functions are very usable still.

We can explore dedicating additional engineering time to isolating the
regions of interest better if these turn out to be a pain point. But in
the meantime lets land these.