congestion: traffic light strategy (#10743)
A congestion strategy featuring 3 levels of traffic control.

This strategy is designed to guarantee:
- Hard limit on incoming queue length measured in attached gas.
- Soft limit on total size of receipts in incoming + outgoing receipts.
- Provably no deadlocks.

Each shard set its own traffic signal to one of:

- green (full-speed)
- amber (senders must stop accepting more load for this shard)
- red (senders must stop all load to this shard)

When a shard turns on the red light, manual traffic control ensures no
deadlocks occur. The congested shard will assign one shard each turn
that is allowed to send a limited amount of traffic. This way, some
amount
of progress is guaranteed on every shard and complete deadlocks are
impossible.