Panic in `CreateSnapshotRequest` handle if snapshot creation failed (#10765)
Currently we have two reasons for snapshot creation: resharding and
dumping of state sync state parts. In both of those cases state snapshot
is a necessary part of the correct node behaviour. So, we should panic
if its creation failed.
In the future we might want to add optional snapshot creation option,
and in this case we will need to add additional check before panicking.

Snapshot creation in regular neard run is handled via
`StateSnapshotActor`.
In some tests and tools we call
`checkpoint_hot_storage_and_cleanup_columns` directly, but right now we
don't have a reason to changing those.