store::trie instrumentation / make otel filtering more flexible (#10919)
This is two changes in one, unfortunately. I don't think I've enough
capacity to split them up cleanly.

Anyway, the more important change here is that the opentelemetry filter
now allows for target-based level overrides. This results in some syntax
changes and such, which users of this functionality should be aware of,
but since you can count the number of them on a hand currently, this is
not a difficult one to push through.)

The other change is just adding some extra instrumentation around some
of the interesting operations within the `near-store`.

cc @bowenwang1996 @vlad-kopcil