feat(mocknet): add a make-backup command (#10931)
It is often useful to stop all the nodes and make backups of their data
dirs so that we can reset to that point later. For example, when testing
resharding it's nice to have a backup at a point right before the
resharding epoch starts so that we can test the resharding part over and
over without having to start from the beginning. So this PR adds a
`make_backup` RPC method to `neard_runner.py` that stops neard and
copies its data dir, remembering the `backup_id` parameter associated
with it. Then in the `reset` command, the same `backup_id` can be
referenced to restore to a previous backup.