runtime: prefetch account keys and data for chunk finalization (#10935)
This significantly reduces the cost of finalize, especially for shard 3
where they were the largest:


![before](https://github.com/near/nearcore/assets/679122/c5b3fb26-3582-43f5-8935-56b8da98acad)

![after](https://github.com/near/nearcore/assets/679122/d3d8ed3c-3423-4a9e-86db-5f8538843a2f)

The effect on the overall apply duration is not as high however:


![image](https://github.com/near/nearcore/assets/679122/4844ad73-b11e-41f0-b1ea-765b1c475b3d)

![image](https://github.com/near/nearcore/assets/679122/548da67d-c761-4adc-99b4-a4a8cc6e408f)

This is understandable, as ultimately all of this data that we were
seeing in `finalize` still needs to be retrieved. The important part,
however is that the duration of this is much more consistent and less
jumpy now.

cc @saketh-are keep an eye out on whether we need to prefetch receipt
data for yield-resume receipts.

---------

Co-authored-by: Bowen Wang <bowen@near.org>
Co-authored-by: Bowen Wang <bowenwang1996@users.noreply.github.com>