feat: effective flat storage rollback (#10946)
Add feature to move flat head back to allow applying range of blocks for
replayability and performance testing. Does not depend on `apply_chunk`,
uses `DBCol::StateChanges` and some hack to account for other keys which
are not yet recorded there. In future releases hack should be removed
because of #10903.

Command: `./neard flat-storage move-flat-head --shard-id 2 --version 3
back --blocks 100`

Note that it doesn't record "visited" deltas to DB. This command is used
for performance testing during which "future" deltas must not exist.

Co-authored-by: Longarithm <the.aleksandr.logunov@gmail.com>