Add metric for recorded storage proof size, also calculate adjusted size based on number of trie removals (#11000)
In https://github.com/near/nearcore/issues/10890 we considered adding
extra 2000 bytes to the storage proof size for every removal operation
on the trie (Idea 1). This PR implements the logic of recording the
number of removals and calculating the adjusted storage proof size. It's
not used in the soft witness size limit for now, as we have to evaluate
how viable of a solution it is. There are concerns that charging 2000
bytes extra would cause the throughput to drop.

To evaluate the impact of adjusting the size calculation like this, the
PR adds three new metrics.
* `near_receipt_recorded_size` - a histogram of how much storage proof
is recorded when processing a receipt. Apart from
https://github.com/near/nearcore/issues/10890, it'll help us estimate
what the hard per-receipt storage proof size limit should be.
* `near_receipt_adjusted_recorded_size` - a histogram of adjusted
storage proof size calculated when processing a receipt
* `near_receipt_adjusted_recorded_size_ratio` - ratio of adjusted size
to non-adjusted size. It'll allow us to evaluate how much the adjustment
affects the final size. The hope is that contracts rarely delete things,
so the effect will be small (a few percent), but it might turn out that
this assumption is false and the adjusted size is e.g 2x higher that the
non-adjusted one. In that case we'd have to reevaluate whether it's a
viable solution.

I'd like to run code from this branch on shadow validation nodes to
gather data from mainnet traffic.