feat(mocknet): use neard fork-network to set up mocknet nodes (#11034)
This adds support in neard_runner.py for nodes whose initial state will
be created via `neard fork-network` instead of by parsing a big records
json file. This will result in much faster startup times.

It will be implemented by detecting what files exist in the
`~/.near/setup` directory. If they contain just `genesis.json` and
`records.json`, we assume we're running an old version of node setup,
and if it looks like it contains a full NEAR home directory, we will
assume it's prepared to be used with `neard fork-network`. The only
thing that changes for `mirror.py` is that a `genesis_time` argument
will be required in this case in order to get a consistent network from
`neard fork-network set-validators` across all the nodes