Feat(eth-wallet-contract): reproducible build (#11011)
Continuing with the work on eth-wallet contract, this PR adds a
reproducible build process. It uses the Docker image [recommended in
near-sdk-rs](https://github.com/near/near-sdk-rs#building-with-reproducible-builds).
Note that the reproducible build process is skipped if the wasm file is
already present and it is still committed to the repo. This is
intentional because it means developers not wanting to worry about the
eth-wallets contract can safely ignore it entirely when setting up their
development environment. However, it does mean that any changes to the
wallet contract implementation will also require locally deleting the
wasm file and then running the build process to generate the new wasm
file. The test that checks the hash of the wasm file will also need to
be updated with whatever the new hash is.

For reviewers: please try deleting the wasm file yourself and running
the build process (e.g. with `cargo test -p near-wallet-contract
--features nightly`) to ensure that it is indeed reproducible across
machines.

This PR also sets the value of `ADDRESS_REGISTRAR_ACCOUNT_ID` equal to
the agreed upon value `address-map.near`. This account exists on both
testnet and mainnet so the same value can be used for both networks. For
now Aurora holds the keys to this account, but once we are sure the
address registrar contract implementation (present in the previous eth
wallet PR) will not change (e.g. after it has been audited) then we will
delete the keys from that account to lock the contract implementation.