congestion: script to reproduce fine-tuning data (#11078)
This script was used to produce the data in
https://docs.google.com/spreadsheets/d/1Vt_-sgMdX1ncYleikYY8uFID_aG9RaqJOqaVMLQ37tQ/edit#gid=0.

It contains the set of model runs that I found informative to come up
with a proposal for the fine-tuned parameters.
Maybe other people can use it to quickly compare their improvement
ideas, I'm sure we haven't found the "best" strategy, yet, if such a
thing exists.

I think it makes sense to store it along the source. But I am also open
to just storing on the drive and not adding it to git repo.