feat: skip processing old state witnesses (#11081)
This PR introduces protection against wasting stateless validator
resources on processing state witness when chunk endorsement is
guaranteed not to have an effect on the chain state. In particular any
chunk with `height_created` that doesn't exceed the height of the last
final block can be safely skipped.

Part of #10565.