[stateless_validation] Introduce new StateWitnessDistributionActor (#11079)
Part 1

Introduce a new StateWitnessDistributionActor for handling distribution
of state witness across the different chunk validators. It's a stand
alone actor in it's own Arbiter for handling async messages wrt. chunk
state witnesses.

As of this PR, we just sequentially send out the state witness to all
the chunk validators.

In the future StateWitnessDistributionActor would be responsible for the
following:
- Splitting state witness into parts using reed solomon erasure encoding
and sending to chunk validators
- Forwarding parts received from chunk producer to other chunk
validators
- Managing all parts received and reconstructing the state witness