fix(mirror): send all chunk's transactions (#11104)
Before this change `mirror run` gets the list of shards to send
transactions for from the tracked_shards config field. This is pretty
unexpected and unwanted, and means that things won't work properly after
a resharding adds more shards than are included in this config field. So
just send all the transactions we see in a given block's chunks.